+++ LoginApp/LoginApp/.settings/org.eclipse.wst.common.component
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
+    <wb-module deploy-name="LoginApp">
+        <wb-resource deploy-path="/" source-path="/WebContent" tag="defaultRootSource"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/java"/>
+        <property name="context-root" value="LoginApp"/>
+        <property name="java-output-path" value="/LoginApp/build/web/WEB-INF/classes"/>
+    </wb-module>
+</project-modules>
