+++ LoginApp/LoginApp/.settings/org.eclipse.wst.jsdt.ui.superType.container
@@ -1,0 +1,1 @@
+org.eclipse.wst.jsdt.launching.baseBrowserLibrary
