+++ LoginApp/LoginApp/nbproject/private/private.properties
@@ -1,0 +1,4 @@
+deploy.ant.properties.file=/home/unroot/.netbeans/12.6/tomcat80.properties
+j2ee.server.home=/home/unroot/Desktop/apache-tomcat-8.5.75
+j2ee.server.instance=tomcat80:home=/home/unroot/Desktop/apache-tomcat-8.5.75
+user.properties.file=/home/unroot/.netbeans/12.6/build.properties
