+++ LoginApp/LoginApp/src/conf/MANIFEST.MF
