--- LoginApp/nbproject/private/private.properties
@@ -1,4 +1,0 @@
-deploy.ant.properties.file=C:\\Users\\Laser\\AppData\\Roaming\\NetBeans\\12.6\\tomcat90.properties
-j2ee.server.home=C:/Tomcat/apache-tomcat-10.0.13
-j2ee.server.instance=tomcat90:home=C:\\Tomcat\\apache-tomcat-10.0.13
-user.properties.file=C:\\Users\\Laser\\AppData\\Roaming\\NetBeans\\12.6\\build.properties
