--- LoginApp/web/WEB-INF/web.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app version="3.1" xmlns="http://xmlns.jcp.org/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd">
-   <servlet>  
-    <description></description>  
-    <display-name>LoginServlet</display-name>  
-    <servlet-name>LoginServlet</servlet-name>  
-    <servlet-class>com.javatpoint.LoginServlet</servlet-class>  
-  </servlet>  
-  <servlet-mapping>  
-    <servlet-name>LoginServlet</servlet-name>  
-    <url-pattern>/LoginServlet</url-pattern>  
-  </servlet-mapping>  
-  <servlet>  
-    <description></description>  
-    <display-name>ProfileServlet</display-name>  
-    <servlet-name>ProfileServlet</servlet-name>  
-    <servlet-class>com.javatpoint.ProfileServlet</servlet-class>  
-  </servlet>  
-  <servlet-mapping>  
-    <servlet-name>ProfileServlet</servlet-name>  
-    <url-pattern>/ProfileServlet</url-pattern>  
-  </servlet-mapping>  
-  <servlet>  
-    <description></description>  
-    <display-name>LogoutServlet</display-name>  
-    <servlet-name>LogoutServlet</servlet-name>  
-    <servlet-class>com.javatpoint.LogoutServlet</servlet-class>  
-  </servlet>  
-  <servlet-mapping>  
-    <servlet-name>LogoutServlet</servlet-name>  
-    <url-pattern>/LogoutServlet</url-pattern>  
-  </servlet-mapping>  
-    <session-config>
-        <session-timeout>
-            30
-        </session-timeout>
-    </session-config>
-</web-app>
